% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_utils.R
\name{compute_pca_embedding_from_tensor}
\alias{compute_pca_embedding_from_tensor}
\title{PCA embedding from two feature matrices (signature x regulator)}
\usage{
compute_pca_embedding_from_tensor(tensor, n_components = 3)
}
\arguments{
\item{tensor}{Lista com:
\itemize{
\item features_sig: matriz N x P
\item features_int: matriz N x P
}}

\item{n_components}{Número de PCs a manter (default = 3)}
}
\value{
Lista com:
\itemize{
\item pca: objeto prcomp
\item coords_sig: N x n_components (scores da assinatura)
\item coords_int: N x n_components (scores do regulador)
}
}
\description{
PCA embedding from two feature matrices (signature x regulator)
}
\keyword{internal}
