% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoCircuitry-package.R, R/aaa.R
\docType{package}
\name{GeoCircuitry-package}
\alias{GeoCircuitry-package}
\title{GeoCircuitry: Geometric Analysis of Circuitry Signatures}
\description{
Provides tools for building, analyzing and visualizing geometric representations of biological circuitry signatures and their regulatory interactions in 3D PCA space. The package enables construction of signature and regulator geometries, computation of convergence metrics, and interactive 3D visualization of circuitry polytopes.

Provides tools for building, analyzing and visualizing geometric representations of biological circuitry signatures and their regulatory interactions in 3D PCA space. The package enables construction of signature and regulator geometries, computation of convergence metrics, and interactive 3D visualization of circuitry polytopes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/seuusuario/GeoCircuitry}
  \item Report bugs at \url{https://github.com/seuusuario/GeoCircuitry/issues}
}


Useful links:
\itemize{
  \item \url{https://github.com/seuusuario/GeoCircuitry}
  \item Report bugs at \url{https://github.com/seuusuario/GeoCircuitry/issues}
}

}
\author{
\strong{Maintainer}: Seu Nome \email{seu.email@instituicao.br} (\href{https://orcid.org/SEU-ORCID-AQUI}{ORCID})

}
\keyword{internal}
