% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoCircuitry-package.R, R/aaa.R
\docType{package}
\name{GeoCircuitry}
\alias{GeoCircuitry}
\title{GeoCircuitry: Geometric Analysis of Circuitry Signatures}
\description{
Provides tools for building, analyzing and visualizing geometric
representations of biological circuitry signatures and their regulatory
interactions in 3D PCA space.

Provides tools for building, analyzing and visualizing geometric
representations of biological circuitry signatures and their regulatory
interactions in 3D PCA space.
}
\details{
The main functions of the package include:
\itemize{
\item \code{\link{build_signature_geometry}} - Build signature geometry
\item \code{\link{build_regulator_geometry}} - Build regulator geometry
\item \code{\link{compute_circuitry_convergence}} - Compute convergence metrics
\item \code{\link{plot_circuitry_3d}} - Interactive 3D visualization
}
}
\references{
Add relevant references here
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/seuusuario/GeoCircuitry}
  \item Report bugs at \url{https://github.com/seuusuario/GeoCircuitry/issues}
}


Useful links:
\itemize{
  \item \url{https://github.com/seuusuario/GeoCircuitry}
  \item Report bugs at \url{https://github.com/seuusuario/GeoCircuitry/issues}
}

}
\author{
Seu Nome \href{mailto:seu@email.com}{seu@email.com}

\strong{Maintainer}: Seu Nome \email{seu.email@instituicao.br} (\href{https://orcid.org/SEU-ORCID-AQUI}{ORCID})

}
