% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circuitry_3d.R
\name{plot_signature_polytope_3d}
\alias{plot_signature_polytope_3d}
\title{3D polytope da ASSINATURA (signature) para um circuitry}
\usage{
plot_signature_polytope_3d(
  poly = NULL,
  sig_geom = NULL,
  reg_geom = NULL,
  index = 1,
  alpha_faces = 0.25,
  size_vertices = 4,
  size_bary = 7
)
}
\arguments{
\item{poly}{Objeto retornado por build_circuitry_polytope().
Alternativamente, você pode passar (sig_geom, reg_geom, index) por compatibilidade.}

\item{sig_geom}{Objeto retornado por build_signature_geometry() (compatibilidade).}

\item{reg_geom}{Objeto retornado por build_regulator_geometry() (compatibilidade).}

\item{index}{Índice do circuitry (compatibilidade).}
}
\value{
plotly
}
\description{
3D polytope da ASSINATURA (signature) para um circuitry
}
