% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeometryCore.R, R/pca_utils.R
\name{project_to_pca}
\alias{project_to_pca}
\title{Internal: PCA embedding from generic tensor (sig + int)}
\usage{
project_to_pca(pca_obj, X, n_components = 3)

project_to_pca(pca_obj, X, n_components = 3)
}
\arguments{
\item{pca_obj}{Objeto prcomp retornado por compute_pca_embedding_from_tensor()}

\item{X}{Matriz N x P com as mesmas colunas usadas no PCA}

\item{n_components}{Número de PCs a retornar}

\item{tensor}{Lista com:
\itemize{
\item features_sig: matriz N x P (assinatura principal)
\item features_int: matriz N x P (regulador)
}}
}
\value{
Lista com:
\itemize{
\item pca: objeto prcomp
\item coords_sig: coordenadas N x n_components para o lado sig
\item coords_int: coordenadas N x n_components para o lado int
}

Matriz N x n_components com novas coordenadas projetadas
}
\description{
Internal: PCA embedding from generic tensor (sig + int)

Project new points into an existing PCA space
}
\keyword{Internal:}
\keyword{PCA}
\keyword{arbitrary}
\keyword{existing}
\keyword{internal}
\keyword{into}
\keyword{matrix}
\keyword{project}
